+++ lib/class_variants/configuration.rb
@@ -1,0 +1,11 @@
+module ClassVariants
+  class Configuration
+    def process_classes_with(&block)
+      if block_given?
+        @process_classes_with = block
+      else
+        @process_classes_with
+      end
+    end
+  end
+end